Fix player ship movement (#245)
Minor fix to #218 and #129 

Now root of the flyable ship (PlayerShip script) is moving with the rest
of the ship.
Changed slightly for compatibility so that the Playership which uses
modules to move works with the ship which is controlled by
CharacterBody3D -> so the shipController is added which is an overlay
for the normal ship to move it freely.

Minor fix do not print text when module button is null