Boss script (#251)
Created two sample bosses prefabs with scripts. Those two can be found
in the prefabs directory.
Bosses are made as enemies - they use state machine.
Both bosses are modular to some extent, to show interesting behaviour of
them.
Boss node can be either miniboss or normal boss. After deafeting the
miniboss the player is returned to map selector, otherwise to credits.

Added also a shootable ship to test the bosses in game.

To shoot press Q.


To test it you can:
- launch the sandbox scene (wierzba/boss_arena/boss_arena), you can even
change the boss to spawn in the DummyBossNodeSetter node.
- launch main_menu scene (the main scene of the game) and go through
bottom missions and kill the boss in the middle of a game XD