Poi arrow (#303)
Created POI prefab and script, which allows displaying both sprite and
arrow on the minimap. If the sprite should be visible on the minimap it
will be, but if too far it will we replaced with arrow, which size
depends od the distance. If too far, even the arrow will be hidden.

To test you can launch the scene:
sandbox/wierzba/poi_arrow/poi_arrow_test.tscn
or even main scene (going from main menu, by pressing play)

Added poi prefabs to enemies, and created portal prefab to test it for
missions