Battery module (#407)
Resolves #358 
The battery increases maximum energy of the ship when it is attached,
and decreases it back, when it is detached.

-----

If you attach the battery in the shop, exit the shop, enter the shop,
detach the battery and exit the shop, the UI breaks - the battery
overlay is displayed to the right from where it should be. I *really*
tried to fix this, but I have no idea how. The energy UI was made ages
ago, so I'd consider this bug as another task tbh.

From the solutions I tried so far, I can see that in the shop (except
the first shop, when starting the game) there are actually *two* HUDs,
one of which is in the detached fly_ship scene. I tried just removing
the HUD in the shop, as in reality it's not that useful for anything
(the minimap is always empty anyways, and the ship preview is not super
necessary), but for some reason the ship builder requires HUD for its
raycast camera. Manually adding another camera is also not an option, as
only the HUD camera can be zoomed in and out. I have no idea how to
untangle this spaghetti without tearing everythin apart and building it
from the ground up, so I will leave this bug as is for now.

-----

After adding energy cost to modules, I noticed that it is also
incorrectly displayed in the energy HUD.