Boss models (#424)
Models and textures of the Juggernaut and Laserstar bosses.

Each boss is split into multiple different models, corresponding to its
different destructible parts.

The triangle count is rather high (50k for Juggernaut and 42k for
Laserstar), but since there will never be more than one boss on the
screen, I think it's acceptable. I can reduce it later if it causes any
problems.

-----

## Juggernaut


![image](https://github.com/user-attachments/assets/d0b19fad-180e-4871-9234-49841b1deca9)


![image](https://github.com/user-attachments/assets/55576b15-f73b-4277-8ed2-e535ecea88c6)

<br>

## Laserstar


![image](https://github.com/user-attachments/assets/5b6d79b7-d145-47b6-b8bf-1ca1ca4933c0)


![image](https://github.com/user-attachments/assets/c5b67d08-8b95-4606-a253-17e9a0f0ea6e)


![image](https://github.com/user-attachments/assets/105cbbf8-6e1f-43bb-9214-90f36fcdd325)