Vortex warning (#456)
Created vortex warning.
It shows that the player is going to fall into vortex (either because he
is close, or with current speed he will fall there in a few seconds).
Added special effect, when player is inside.

To test it, just turn on the game and on the first map go to the vortex

---------

Co-authored-by: Lenetis <70149832+Lenetis@users.noreply.github.com>