Reset sector map (#490)
Properly reset sector map on a start of a game.

Everything resetable is now reseted when the game starts.

To test it as quickly as possible, in the scene
"basic_movement_tutorial.tscn" move the shop_miniature much closer.
To test just play the game and finish at least two escape missions.
Return to main menu, and play again to see that it properly resets.