Roerder UI elements (#494)
Vortex warning is now at the bottom of the UI elements order.
So the main menu and dev button (to shop) is now clickable.

To test just finish tutorial map (or move the shop_miniature close
enough), leave the shop, and then try intercting with pause menu and dev
button