Ship sounds (#504)
Ship sounds update!
(Enemy and UI sounds are still TODO)

- Reworked SoundEmitter slightly, to make more effects possible
- Added activation sounds to all modules
- Added portal jump sounds
- Added module destroyed sounds
- Added collision sounds
    - As a byproduct, also collision particles ^^
- Added damage taken sounds
- As a byproduct, reworked Damage and DamageController, to be able to
differentiate damage types (normal, collision, vortex, beam)

-----

Collision and module destroyed sounds are emitted by their respective
particles. This looks weird, but that was the easiest way to implement
it that I could think of.

The damage taken sounds work and behave correctly, but their
implementation is a bit spaghetti. I wouldn't be too concerned about it,
but it might be worth considering to place them in every projectile
instead (played on projectile's death).