Vortex boss new design (#511)
Redesigned vortex boss. It consists of tentacles which blocks weapons
and can be destroyed by destroing two orbs laying on it. If all
tentacles are destroyed the ring protecting the core is destroyed.

Attack are still the same. Changed models for new version, and added
pore model for releasing the cloud (the cloud is still missing)

To test try playing the scene:
```
res://sandbox/wierzba/boss_arena/boss_arena.tscn
```

You should turn the invincibility (F3) to be able to kill with default
ship (it not too manevuarable for this boss)

for even faster testing just set layer in damageable in vortex_core.tscn
to 2 and 3 and mask to 2 and 3 - the boss will damage self