Fix enemies shooting the player ship when it's moving down through a portal (#530)
When the player ship was playing the jump animation and moving to the
next level, the enemies were able to rotate downwards towards it and
shoot it anyways. This is fixed by setting vector.y to 0 in all enemy
states that deal with player position. Some of the changes might be
unnecessary, but since we never want to include the Y height in our
calculations, this shouldn't break anything