Spaghetti fix for bosses and composite states
Composite states weren't working when one of the states was setting velocity and the other was not (because the other one was clearing velocity in that case). The fix is to disable clearing velocity for substates in composite states (only the main composite state now clears velocity, not its substates). Not pretty, but it works

Also, the rotations being in state_update caused some weird bugs, they were resolved by moving them to state_physics_update