Build ship ui tutorial better placement (#558)
For build cutscene the arrow was made bigger,
for inventory cutscene arrow was moved to the right side, and whole text
box was moved slightly to the right.

Moreover, camera zoom in shop is not dependent on the camere zoom in the
fly phase. Camera in shop ignores settings, and won't change the zoom in
the fly phase. Thanks to that, we can make sure that everything is
visible during tutorial.

For tester:
- check if both arrows in tutorials point to correct place
- check if camera work properly